+++ packages/ocean-react/src/_shared/components/InternalListActions/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useSwipeGesture } from './useSwipeGesture';
+export { default as useClickOutside } from './useClickOutside';