+++ packages/ocean-react/src/index.ts
@@ -52,6 +52,9 @@ export * from './Link';
 export { default as TextArea } from './TextArea';
 export * from './TextArea';
 
+export { default as ListSelectable } from './ListSelectable';
+export * from './ListSelectable';
+
 export { default as TextListItem } from './TextListItem';
 export * from './TextListItem';
 
@@ -122,3 +125,6 @@ export * from './WebNotification';
 
 export { default as SettingsListItem } from './SettingsListItem';
 export * from './SettingsListItem';
+
+export { default as CardListAction } from './CardListAction';
+export * from './CardListAction';